+++ Realm/RLMArrayLinkView.mm
@@ -36,7 +36,7 @@
 + (RLMArrayLinkView *)arrayWithObjectClassName:(NSString *)objectClassName
                                           view:(tightdb::LinkViewRef)view
                                          realm:(RLMRealm *)realm {
-    RLMArrayLinkView *ar = [[RLMArrayLinkView alloc] initViewWithObjectClassName:objectClassName];
+    RLMArrayLinkView *ar = [[RLMArrayLinkView alloc] initWithObjectClassName:objectClassName standalone:NO];
     ar->_backingLinkView = view;
     ar->_realm = realm;
     return ar;
@@ -185,13 +185,6 @@ static inline void RLMValidateObjectClass(RLMObject *obj, NSString *expected) {
     _backingLinkView->set(index, object->_row.get_index());
 }
 
-- (NSString *)JSONString {
-    RLMLinkViewArrayValidateAttached(self);
-
-    @throw [NSException exceptionWithName:@"RLMNotImplementedException"
-                                   reason:@"Not yet implemented" userInfo:nil];
-}
-
 - (NSUInteger)indexOfObject:(RLMObject *)object {
     // check attached for table and object
     RLMLinkViewArrayValidateAttached(self);
@@ -228,19 +221,16 @@ static inline void RLMValidateObjectClass(RLMObject *obj, NSString *expected) {
     self->_backingLinkView->remove_all_target_rows();
 }
 
-- (RLMArray *)arraySortedByProperty:(NSString *)property ascending:(BOOL)ascending
+- (RLMResults *)sortedResultsUsingDescriptors:(NSArray *)properties
 {
     RLMLinkViewArrayValidateAttached(self);
-    
-    // validate sort criteria
-    RLMProperty *prop = RLMValidatedPropertyForSort(_realm.schema[self.objectClassName], property);
-    
-    // apply order
-    tightdb::TableView const &tv = _backingLinkView->get_sorted_view(prop.column, ascending);
-    RLMArrayTableView *ar = [RLMArrayTableView arrayWithObjectClassName:self.objectClassName
-                                                                   view:tv
-                                                                  realm:_realm];
-    return ar;
+
+    std::vector<size_t> columns;
+    std::vector<bool> order;
+    RLMGetColumnIndices(_realm.schema[_objectClassName], properties, columns, order);
+
+    tightdb::TableView const &tv = _backingLinkView->get_sorted_view(move(columns), move(order));
+    return [RLMResults resultsWithObjectClassName:self.objectClassName view:tv realm:_realm];
 }
 
 @end